requirements.txt
python-telegram-bot==20.7
groq==0.9.0
sentence-transformers==2.2.2
python-dotenv==1.0.0